Hide bounded/unbounded implementation details from grid tilers
