+++ x-pack/plugin/spatial/src/test/java/org/elasticsearch/xpack/spatial/search/aggregations/bucket/geogrid/GeoHexTilerTests.java
@@ -28,15 +28,11 @@ import java.util.Arrays;
 import static org.elasticsearch.xpack.spatial.util.GeoTestUtils.geoShapeValue;
 import static org.hamcrest.Matchers.equalTo;
 
-public class GeoHexTilerTests extends GeoGridTilerTestCase {
-    @Override
-    protected GeoGridTiler getUnboundedGridTiler(int precision) {
-        return new UnboundedGeoHexGridTiler(precision);
-    }
+public class GeoHexTilerTests extends GeoGridTilerTestCase<GeoHexGridTiler> {
 
     @Override
-    protected GeoGridTiler getBoundedGridTiler(GeoBoundingBox bbox, int precision) {
-        return new BoundedGeoHexGridTiler(precision, bbox);
+    protected GeoHexGridTiler getGridTiler(GeoBoundingBox bbox, int precision) {
+        return GeoHexGridTiler.makeGridTiler(precision, bbox);
     }
 
     @Override
@@ -57,7 +53,7 @@ public class GeoHexTilerTests extends GeoGridTilerTestCase {
 
     @Override
     protected long getCellsForDiffPrecision(int precisionDiff) {
-        return UnboundedGeoHexGridTiler.calcMaxAddresses(precisionDiff);
+        return GeoHexGridTiler.calcMaxAddresses(precisionDiff);
     }
 
     public void testLargeShape() throws Exception {
@@ -73,7 +69,7 @@ public class GeoHexTilerTests extends GeoGridTilerTestCase {
         for (int precision = 0; precision < 4; precision++) {
             GeoShapeCellValues values = new GeoShapeCellValues(
                 makeGeoShapeValues(value),
-                getBoundedGridTiler(boundingBox, precision),
+                getGridTiler(boundingBox, precision),
                 NOOP_BREAKER
             );
             assertTrue(values.advanceExact(0));
@@ -107,7 +103,7 @@ public class GeoHexTilerTests extends GeoGridTilerTestCase {
             String msg = "Failed " + WellKnownText.toWKT(point) + " at resolution " + res + " with precision " + precision;
             GeoShapeCellValues values = new GeoShapeCellValues(
                 makeGeoShapeValues(value),
-                getBoundedGridTiler(boundingBox, precision),
+                getGridTiler(boundingBox, precision),
                 NOOP_BREAKER
             );
             assertTrue(values.advanceExact(0));
@@ -133,7 +129,7 @@ public class GeoHexTilerTests extends GeoGridTilerTestCase {
         GeoShapeValues.GeoShapeValue value = geoShapeValue(geometry);
         GeoShapeCellValues cellValues = new GeoShapeCellValues(
             makeGeoShapeValues(value),
-            getBoundedGridTiler(geoBoundingBox, precision),
+            getGridTiler(geoBoundingBox, precision),
             NOOP_BREAKER
         );
 
@@ -151,11 +147,7 @@ public class GeoHexTilerTests extends GeoGridTilerTestCase {
             1.7481549674935762E-110 -75.113250736563, 1.7481549674935762E-110 -90.0))""";
         Geometry geometry = WellKnownText.fromWKT(StandardValidator.instance(true), true, polygon);
         GeoShapeValues.GeoShapeValue value = geoShapeValue(geometry);
-        GeoShapeCellValues unboundedCellValues = new GeoShapeCellValues(
-            makeGeoShapeValues(value),
-            getUnboundedGridTiler(precision),
-            NOOP_BREAKER
-        );
+        GeoShapeCellValues unboundedCellValues = new GeoShapeCellValues(makeGeoShapeValues(value), getGridTiler(precision), NOOP_BREAKER);
 
         assertTrue(unboundedCellValues.advanceExact(0));
         int numBuckets = unboundedCellValues.docValueCount();
@@ -171,11 +163,7 @@ public class GeoHexTilerTests extends GeoGridTilerTestCase {
             180.0 90.0, 36.98661841690625 90.0, 36.98661841690625 69.44049730644747))""";
         Geometry geometry = WellKnownText.fromWKT(StandardValidator.instance(true), true, polygon);
         GeoShapeValues.GeoShapeValue value = geoShapeValue(geometry);
-        GeoShapeCellValues unboundedCellValues = new GeoShapeCellValues(
-            makeGeoShapeValues(value),
-            getUnboundedGridTiler(precision),
-            NOOP_BREAKER
-        );
+        GeoShapeCellValues unboundedCellValues = new GeoShapeCellValues(makeGeoShapeValues(value), getGridTiler(precision), NOOP_BREAKER);
 
         assertTrue(unboundedCellValues.advanceExact(0));
         int numBuckets = unboundedCellValues.docValueCount();
@@ -188,11 +176,7 @@ public class GeoHexTilerTests extends GeoGridTilerTestCase {
         String polygon = "BBOX (-84.24596376729815, 43.36113427778119, 90.0, 83.51476833522361)";
         Geometry geometry = WellKnownText.fromWKT(StandardValidator.instance(true), true, polygon);
         GeoShapeValues.GeoShapeValue value = geoShapeValue(geometry);
-        GeoShapeCellValues unboundedCellValues = new GeoShapeCellValues(
-            makeGeoShapeValues(value),
-            getUnboundedGridTiler(precision),
-            NOOP_BREAKER
-        );
+        GeoShapeCellValues unboundedCellValues = new GeoShapeCellValues(makeGeoShapeValues(value), getGridTiler(precision), NOOP_BREAKER);
 
         assertTrue(unboundedCellValues.advanceExact(0));
         int numBuckets = unboundedCellValues.docValueCount();
@@ -215,7 +199,7 @@ public class GeoHexTilerTests extends GeoGridTilerTestCase {
         GeoShapeValues.GeoShapeValue value = geoShapeValue(geometry);
         GeoShapeCellValues cellValues = new GeoShapeCellValues(
             makeGeoShapeValues(value),
-            getBoundedGridTiler(boundingBox, precision),
+            getGridTiler(boundingBox, precision),
             NOOP_BREAKER
         );
 
@@ -236,7 +220,7 @@ public class GeoHexTilerTests extends GeoGridTilerTestCase {
         GeoShapeValues.GeoShapeValue value = geoShapeValue(geometry);
         GeoShapeCellValues cellValues = new GeoShapeCellValues(
             makeGeoShapeValues(value),
-            getBoundedGridTiler(boundingBox, precision),
+            getGridTiler(boundingBox, precision),
             NOOP_BREAKER
         );
 
@@ -248,11 +232,7 @@ public class GeoHexTilerTests extends GeoGridTilerTestCase {
 
     private void assertCorner(long[] h3bins, Point point, int precision, String msg) throws IOException {
         GeoShapeValues.GeoShapeValue cornerValue = geoShapeValue(point);
-        GeoShapeCellValues cornerValues = new GeoShapeCellValues(
-            makeGeoShapeValues(cornerValue),
-            getUnboundedGridTiler(precision),
-            NOOP_BREAKER
-        );
+        GeoShapeCellValues cornerValues = new GeoShapeCellValues(makeGeoShapeValues(cornerValue), getGridTiler(precision), NOOP_BREAKER);
         assertTrue(cornerValues.advanceExact(0));
         long[] h3binsCorner = ArrayUtil.copyOfSubArray(cornerValues.getValues(), 0, cornerValues.docValueCount());
         for (long corner : h3binsCorner) {
@@ -263,7 +243,7 @@ public class GeoHexTilerTests extends GeoGridTilerTestCase {
     @Override
     protected void assertSetValuesBruteAndRecursive(Geometry geometry) throws Exception {
         int precision = randomIntBetween(1, 4);
-        UnboundedGeoHexGridTiler tiler = new UnboundedGeoHexGridTiler(precision);
+        GeoHexGridTiler tiler = getGridTiler(precision);
         GeoShapeValues.GeoShapeValue value = geoShapeValue(geometry);
 
         GeoShapeCellValues recursiveValues = new GeoShapeCellValues(null, tiler, NOOP_BREAKER);
@@ -284,7 +264,7 @@ public class GeoHexTilerTests extends GeoGridTilerTestCase {
     }
 
     private int addBruteForce(
-        AbstractGeoHexGridTiler tiler,
+        GeoHexGridTiler tiler,
         GeoShapeCellValues values,
         GeoShapeValues.GeoShapeValue geoValue,
         long h3,
@@ -306,15 +286,15 @@ public class GeoHexTilerTests extends GeoGridTilerTestCase {
 
     @Override
     protected int expectedBuckets(GeoShapeValues.GeoShapeValue geoValue, int precision, GeoBoundingBox bbox) throws Exception {
-        BoundedGeoHexGridTiler bounded = bbox == null ? null : new BoundedGeoHexGridTiler(precision, bbox);
-        UnboundedGeoHexGridTiler predicate = new UnboundedGeoHexGridTiler(precision);
+        GeoHexGridTiler bounded = bbox == null ? null : getGridTiler(bbox, precision);
+        GeoHexGridTiler predicate = getGridTiler(precision);
         return computeBuckets(H3.getLongRes0Cells(), bounded, predicate, geoValue, precision);
     }
 
     private int computeBuckets(
         long[] children,
-        BoundedGeoHexGridTiler bounded,
-        UnboundedGeoHexGridTiler predicate,
+        GeoHexGridTiler bounded,
+        GeoHexGridTiler predicate,
         GeoShapeValues.GeoShapeValue geoValue,
         int finalPrecision
     ) throws IOException {
@@ -331,12 +311,8 @@ public class GeoHexTilerTests extends GeoGridTilerTestCase {
         return count;
     }
 
-    private boolean intersects(
-        long h3,
-        GeoShapeValues.GeoShapeValue geoValue,
-        BoundedGeoHexGridTiler bounded,
-        UnboundedGeoHexGridTiler predicate
-    ) throws IOException {
+    private boolean intersects(long h3, GeoShapeValues.GeoShapeValue geoValue, GeoHexGridTiler bounded, GeoHexGridTiler predicate)
+        throws IOException {
         if (bounded != null && bounded.h3IntersectsBounds(h3) == false) {
             return false;
         }
